+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [5.0.4] - 2021-02-05
 ### Changed
-- mysql_class.Master.connect:  Added silent option connect method call.
+- mysql_class.MasterRep.connect, mysql_class.SlaveRep.connect:  Added silent option connect method call.
 
 
 ## [5.0.3] - 2021-01-07